# viewer-svelte/static/openapi.yaml
openapi: 3.0.3
info:
  title: Itinerator API
  description: |
    A modern travel itinerary management system with AI-powered trip planning.

    ## Authentication

    All endpoints require authentication via API key in the `X-API-Key` header.
    Some endpoints also accept `X-OpenRouter-API-Key` for AI features.

    ## Base URL

    - Local Development: `http://localhost:5176/api/v1`
    - Production: `https://your-domain.vercel.app/api/v1`

    ## Streaming Endpoints

    Endpoints marked with `[SSE]` return Server-Sent Events streams.

  version: 1.0.0
  contact:
    name: Itinerator Support
    url: https://github.com/bobmatnyc/itinerator
servers:
  - url: http://localhost:5176/api/v1
    description: Local development
  - url: /api/v1
    description: Current environment
...
security:
  - ApiKeyAuth: []
paths:
...
  /designer/sessions:
    post:
      tags: [Trip Designer]
      summary: Create chat session
      description: Create a new AI chat session for trip planning or help
      operationId: createSession
      parameters:
        - name: X-OpenRouter-API-Key
          in: header
          description: OpenRouter API key (overrides env var)
          required: false
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                itineraryId:
                  type: string
                  format: uuid
                  description: Itinerary ID (required for trip-designer mode)
                mode:
                  type: string
                  enum: [trip-designer, help]
                  default: trip-designer
                  description: Chat mode
      responses:
        '201':
          description: Session created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  sessionId:
                    type: string
                    format: uuid
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
  /designer/sessions/{sessionId}:
    parameters:
      - name: sessionId
        in: path
        required: true
        schema:
          type: string
          format: uuid

    get:
      tags: [Trip Designer]
      summary: Get session details
      description: Get chat session details and message history
      operationId: getSession
      parameters:
        - name: X-OpenRouter-API-Key
          in: header
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Session details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TripDesignerSession'
        '404':
          $ref: '#/components/responses/NotFound'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'

    delete:
      tags: [Trip Designer]
      summary: Delete session
      description: Delete a chat session (used when switching itineraries or cleanup)
      operationId: deleteSession
      parameters:
        - name: X-OpenRouter-API-Key
          in: header
          required: false
          schema:
            type: string
      responses:
        '204':
          description: Session deleted successfully
        '500':
          $ref: '#/components/responses/InternalError'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
  /designer/sessions/{sessionId}/messages:
    parameters:
      - name: sessionId
        in: path
        required: true
        schema:
          type: string
          format: uuid

    post:
      tags: [Trip Designer]
      summary: Send message (non-streaming)
      description: Send a message to the AI and get a complete response
      operationId: sendMessage
      parameters:
        - name: X-OpenRouter-API-Key
          in: header
          required: false
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - message
              properties:
                message:
                  type: string
                  description: User message text
      responses:
        '200':
          description: AI response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '402':
          description: Cost limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
  /designer/sessions/{sessionId}/messages/stream:
    parameters:
      - name: sessionId
        in: path
        required: true
        schema:
          type: string
          format: uuid

    post:
      tags: [Trip Designer]
      summary: Send message (streaming) [SSE]
      description: |
        Send a message to the AI and receive a streaming Server-Sent Events response.

        **Event Types:**
        - `connected` - Initial connection confirmation
        - `text` - Text chunk from AI response
        - `tool_call` - AI is calling a tool (e.g., add_segment)
        - `tool_result` - Tool execution result
        - `structured_questions` - AI asking structured questions
        - `done` - Response complete (includes tokens and cost)
        - `error` - Error occurred
      operationId: sendMessageStream
      parameters:
        - name: X-OpenRouter-API-Key
          in: header
          required: false
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - message
              properties:
                message:
                  type: string
                  description: User message text
      responses:
        '200':
          description: Server-Sent Events stream
          content:
            text/event-stream:
              schema:
                type: string
                description: SSE stream with events (connected, text, tool_call, tool_result, structured_questions, done, error)
              examples:
                text:
                  summary: Text event
                  value: |
                    event: text
                    data: {"content":"Let me help you with that..."}

                tool_call:
                  summary: Tool call event
                  value: |
                    event: tool_call
                    data: {"name":"add_segment","arguments":{"type":"FLIGHT",...}}

                done:
                  summary: Done event
                  value: |
                    event: done
                    data: {"itineraryUpdated":true,"segmentsModified":["seg-123"],"tokens":1500,"cost":0.0025}
        '400':
          $ref: '#/components/responses/BadRequest'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'

  # ==========================================
  # Import
  # ==========================================
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for authentication
...
  schemas:
...
    TripDesignerSession:
      type: object
      required:
        - sessionId
        - mode
        - createdAt
        - messages
      properties:
        sessionId:
          type: string
          format: uuid
        itineraryId:
          type: string
          format: uuid
        mode:
          type: string
          enum: [trip-designer, help]
        createdAt:
          type: string
          format: date-time
        messages:
          type: array
          items:
            $ref: '#/components/schemas/ChatMessage'
        totalTokens:
          type: integer
        totalCost:
          type: number
    ChatMessage:
      type: object
      required:
        - role
        - content
      properties:
        role:
          type: string
          enum: [user, assistant, system]
        content:
          type: string
        timestamp:
          type: string
          format: date-time
        tokens:
          type: integer
        cost:
          type: number
    AgentResponse:
      type: object
      required:
        - response
      properties:
        response:
          type: string
          description: AI response text
        structuredQuestions:
          type: array
          items:
            $ref: '#/components/schemas/StructuredQuestion'
          description: Questions asking for specific information
        itineraryUpdated:
          type: boolean
          description: Whether the itinerary was modified
        segmentsModified:
          type: array
          items:
            type: string
            format: uuid
          description: IDs of segments that were added/modified
        tokens:
          type: integer
          description: Tokens used in this response
        cost:
          type: number
          description: Cost of this response in USD
    StructuredQuestion:
      type: object
      required:
        - id
        - question
        - type
      properties:
        id:
          type: string
          description: Unique question ID
        question:
          type: string
          description: Question text
        type:
          type: string
          enum: [text, date, number, choice, multi-choice]
        options:
          type: array
          items:
            type: string
          description: Available options for choice/multi-choice questions
        required:
          type: boolean
          default: true

    # ==========================================
    # Error Schemas
    # ==========================================
    Error:
      type: object
      required:
        - message
      properties:
        message:
          type: string
          description: Error message
        code:
          type: string
          description: Error code
        details:
          type: object
          additionalProperties: true
          description: Additional error details
  responses:
    BadRequest:
      description: Bad request - invalid input
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    ServiceUnavailable:
      description: Service unavailable (e.g., Trip Designer disabled - no API key)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'